Changes on request and response model
