+++ covalent_ui/api/v1/models/graph_model.py
@@ -15,40 +15,16 @@
 # Covalent is distributed in the hope that it will be useful, but WITHOUT
 # ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
 # FITNESS FOR A PARTICULAR PURPOSE. See the License for more details.
+#
 # Relief from the License may be granted by purchasing a commercial license.
 
-"""Electron Dependency Request and Respone Model"""
-
-from typing import List
+"""Graph request and response model"""
 
 from pydantic import BaseModel
 
-from covalent_ui.api.v1.utils.status import Status
 
+class GraphResponse(BaseModel):
+    """Graph Response Model"""
 
-class LinkModule(BaseModel):
-    """Link Module Validation"""
-
-    id: int
-    electron_id: int
-    parent_electron_id: int
-    edge_name: str
-    parameter_type: str
-    created_at: str
-
-
-class NodeModule(BaseModel):
-    """Node Module Validation"""
-
-    id: int
-    name: str
-    start_time: str
-    end_time: str
-    status: Status
-
-
-class GraphResponseModel(BaseModel):
-    """Graph Response Validation"""
-
-    node: List[NodeModule]
-    links: List[LinkModule]
+    dispatch_id: str = None
+    graph: dict = None
