+++ covalent_ui/api/v1/routes/end_points/electron_routes.py
@@ -27,9 +27,9 @@ from sqlalchemy.orm import Session
 from covalent._results_manager.results_manager import get_result
 from covalent_ui.api.v1.data_layer.electron_dal import Electrons
 from covalent_ui.api.v1.database.config.db import engine
-from covalent_ui.api.v1.models.electrons_model import FileOutput
-from covalent_ui.api.v1.models.lattices_model import (
+from covalent_ui.api.v1.models.electrons_model import (
     ElectronExecutorResponse,
+    ElectronFileOutput,
     ElectronFileResponse,
     ElectronResponse,
 )
@@ -99,7 +99,7 @@ def get_electron_inputs(dispatch_id: uuid.UUID, electron_id: int) -> str:
 
 
 @routes.get("/{dispatch_id}/electron/{electron_id}/details/{name}")
-def get_electron_file(dispatch_id: uuid.UUID, electron_id: int, name: FileOutput):
+def get_electron_file(dispatch_id: uuid.UUID, electron_id: int, name: ElectronFileOutput):
     """
     Get Electron details
     Args:
